requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
openai>=1.14.0
pyyaml>=6.0.1
httpx>=0.27.0
aiofiles>=23.0.0
